requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
redis==5.0.1
gunicorn==21.2.0
flup6==1.1.1
whitenoise==6.6.0
python-dotenv==1.0.0
django-environ==0.11.2
requests==2.31.0
openai==1.3.5
Pillow==10.1.0
pdf2image==1.16.3
pytesseract==0.3.10
numpy==1.26.2
pandas==2.1.3
python-magic==0.4.27

# Security
django-axes==6.1.1
django-csp==3.7
django-ratelimit==4.1.0
django-secure==1.0.2

# Monitoring
psutil==5.9.6
hiredis==2.2.3
django-redis==5.4.0

# Testing and Development
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.0 